+++ docs/changelog.rst
@@ -9,6 +9,9 @@ CHANGELOG
 
 * Add possibility to show geometries overprinted on topologies (#105)
 * Add possibility to create attachment with external link
+* Add command import hydrobiologie stations hubeau
+* Upgrade api hubeau PC quality
+
 
 1.0.3 (2022-12-15)
 -------------------------